Added GraphQL ResponseStatusExceptions for AccountGraphQlController

+++ eborrow/src/main/java/de/fhdo/eborrow/graphqlapi/AccountGraphQlController.java
@@ -7,36 +7,61 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.server.ResponseStatusException;
 
 @Controller
 public class AccountGraphQlController {
 	private final AccountService accountService;
 
 	@Autowired
-	public AccountGraphQlController(AccountService accountService) {this.accountService = accountService;}
+	public AccountGraphQlController(AccountService accountService) {
+		this.accountService = accountService;
+	}
 
 	@QueryMapping("accounts")
 	public Iterable<AccountDto> getAllAccounts() {
 		return accountService.getAccounts();
 	}
 
-	// TODO Zak: Fehlerstatus uebertragen?
 	@QueryMapping("account")
-	public RichAccountDto getAccountById(@Argument Long id) {
-		return accountService.getRichAccountById(id);
+	public RichAccountDto getRichAccountById(@Argument Long id) {
+		RichAccountDto richAccountDto = accountService.getRichAccountById(id);
+		if (richAccountDto == null) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+
+		return richAccountDto;
 	}
 
 	@MutationMapping("createAccount")
 	public AccountDto createAccount(@Argument AccountDto account) {
-		Long accountId = accountService.addAccount(account);
+		if (account == null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
 
-		return accountService.getAccountById(accountId);
+		Long accountId = accountService.addAccount(account);
+		if (accountId == null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
+
+		AccountDto result = accountService.getAccountById(accountId);
+
+		return result;
 	}
 
 	@MutationMapping("updateAccount")
 	public AccountDto updateAccount(@Argument AccountDto account) {
+		if (account == null || account.getId() == null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
+
 		boolean succeeded = accountService.updateAccount(account);
+		if (!succeeded) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
+
 		AccountDto result = accountService.getAccountById(account.getId());
 
 		return result;
@@ -44,12 +69,26 @@ public class AccountGraphQlController {
 
 	@MutationMapping("deleteAccount")
 	public boolean deleteAccount(@Argument Long id) {
-		return accountService.deleteAccount(id);
+		if (id == null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID is required");
+		}
+
+		boolean succeeded = accountService.deleteAccount(id);
+
+		return succeeded;
 	}
 
 	@MutationMapping("addGameToAccount")
 	public RichAccountDto addGameToAccount(@Argument Long accountId, @Argument Long gameId) {
+		if (accountId == null || gameId == null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID is required");
+		}
+
 		boolean succeeded = accountService.addGameToAccount(accountId, gameId);
+		if (!succeeded) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
+
 		RichAccountDto result = accountService.getRichAccountById(accountId);
 
 		return result;
@@ -57,7 +96,15 @@ public class AccountGraphQlController {
 
 	@MutationMapping("unlistGameFromAccount")
 	public RichAccountDto unlistGameFromAccount(@Argument Long accountId, @Argument Long gameId) {
+		if (accountId == null || gameId == null) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID is required");
+		}
+
 		boolean succeeded = accountService.unlistGameFromAccount(accountId, gameId);
+		if (!succeeded) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		}
+
 		RichAccountDto result = accountService.getRichAccountById(accountId);
 
 		return result;
